Small updates testing F

+++ tools.py
@@ -88,7 +88,7 @@ def vlen(x):
 def sqc(x):
     S = np.array([[    0, -x[2],  x[1]],
                   [ x[2],     0, -x[0]],
-                  [-x[1],  x[0],     0]])
+                  [-x[1],  x[0],     0]], dtype='float64')
     return S
